Finished DebugRenderer. Improved VisualEffect classes' design by using std::enable_share_from_this.

+++ include/FalconEngine/Graphics/Renderer/Resource/BufferCircular.h
@@ -22,10 +22,10 @@ protected:
 
 private:
     // Where the data offset should be when the next fill phrase begins.
-    int64_t mBufferDataOffsetNext;
+    size_t mBufferDataOffsetNext;
 
     // How much data should be keep safely at once without wrapping the data offset.
-    size_t  mBufferZoneSize;
+    size_t mBufferZoneSize;
 };
 
 }
